refactor: 파일 구조 정리 및 not-found 페이지 버튼 로직 추가
