+++ src/app/not-found.tsx
@@ -1,6 +1,10 @@
+'use client';
+
 import { Button } from '@components/ui/button';
+import { useRouter } from 'next/navigation';
 
 export default function NotFoundPage() {
+  const router = useRouter();
   return (
     <main className="flex min-h-screen flex-col items-center justify-center gap-y-4 p-4">
       <img src="/icons/404.png" alt="404 이미지" className="mb-8 h-auto w-full max-w-[200px]"></img>
@@ -10,7 +14,7 @@ export default function NotFoundPage() {
         요청하신 페이지가 사라졌거나 <br />
         잘못된 경로입니다.
       </p>
-      <Button variant="outline" className="text-md rounded-3xl">
+      <Button variant="outline" onClick={() => router.back()} className="text-md rounded-3xl">
         이전 화면으로 돌아가기
       </Button>
     </main>
